64 bit support fixes
+ In several places pointers were assumed to have a size of 4 byte (and thus
. we could squeeze 'em in an unsigned int), of course this is wrong for
. 64 bit systems. Fixed where needed.
+ Added -fPIC to CFLAGS_BASE, this is needed as GCC told me...

Tim


git-svn-id: https://lechuck.informatik.rwth-aachen.de:2342/svn/fawkes/trunk@120 fc82b16a-671c-0410-8732-9903a8d2dd0d

+++ src/firevision/models/color/lookuptable.cpp
@@ -382,7 +382,7 @@ ColorModelLookupTable::composeFilename(const string format)
   struct utsname uname_info;
   uname( &uname_info );
 
-  unsigned int loc = rv.find( "%h" );
+  size_t loc = rv.find( "%h" );
   while (loc != string::npos) {
     rv.replace( loc, 2, uname_info.nodename );
     loc = rv.find( "%h" );
